+++ src/app/header/header.component.html
@@ -15,7 +15,7 @@
 <div class="access-links">
   <ng-container *ngIf="isLoggedIn(); else anonymous">
     <a class="clickable-title">{{ getLoggedInUserName() }}</a> |
-    <a class="clickable-title" routerLink="logout">Logout</a>
+    <a class="clickable-title" (click)="logout()">Logout</a>
   </ng-container>
   <ng-template #anonymous>
     <a class="clickable-title" routerLink="login">Login</a> |
