+++ src/app/modules/login/login.component.html
@@ -22,5 +22,8 @@
     >
       Login
     </button>
+    <div *ngIf="errorMessage" class="error-message">
+      {{errorMessage}}
+    </div>
   </form>
 </div>
