+++ src/app/modules/sign-up/sign-up.component.scss
@@ -46,11 +46,6 @@ input[type="password"] {
   }
 }
 
-.error-message {
-  color: #e20606;
-  font-size: 14px;
-}
-
 .buttons-container {
   display: flex;
   justify-content: flex-end;
