+++ src/styles.scss
@@ -25,3 +25,8 @@ button {
     border: none;
   }
 }
+
+.error-message {
+  color: #e20606;
+  font-size: 14px;
+}
